refactored urn history using a baseclass, moved vulcanize implementation to its own file
