+++ tests/manual_test_urn_history.py
@@ -22,7 +22,8 @@ import sys
 from datetime import datetime, timedelta
 from web3 import Web3, HTTPProvider
 
-from auction_keeper.urn_history import UrnHistory
+from auction_keeper.urn_history import ChainUrnHistoryProvider
+from auction_keeper.urn_history_vulcanize import VulcanizeUrnHistoryProvider
 from pymaker.deployment import DssDeployment
 
 
@@ -45,7 +46,7 @@ from_block = int(sys.argv[2]) if len(sys.argv) > 2 else 8928152
 # Retrieve data from chain
 started = datetime.now()
 print(f"Connecting to {sys.argv[1]}...")
-uh = UrnHistory(web3, mcd, ilk, from_block, None, None)
+uh = ChainUrnHistoryProvider(web3, mcd, ilk, from_block)
 urns_logs = uh.get_urns()
 elapsed: timedelta = datetime.now() - started
 print(f"Found {len(urns_logs)} urns from block {from_block} in {elapsed.seconds} seconds")
@@ -54,7 +55,7 @@ print(f"Found {len(urns_logs)} urns from block {from_block} in {elapsed.seconds}
 # Retrieve data from Vulcanize
 started = datetime.now()
 print(f"Connecting to {vulcanize_endpoint}...")
-uh = UrnHistory(web3, mcd, ilk, None, vulcanize_endpoint, vulcanize_key)
+uh = VulcanizeUrnHistoryProvider(mcd, ilk, vulcanize_endpoint, vulcanize_key)
 urns_vdb = uh.get_urns()
 elapsed: timedelta = datetime.now() - started
 print(f"Found {len(urns_vdb)} urns from Vulcanize in {elapsed.seconds} seconds")
